requirements.txt
asgiref==3.7.2
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
geoip2==2.9.0
gunicorn==21.2.0
idna==3.10
maxminddb==2.6.2
numpy==2.1.2
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.3
